+++ integration/deltaspike-data/base/src/main/java/com/blazebit/persistence/deltaspike/data/base/handler/KeysetAwareSliceImpl.java
@@ -92,7 +92,7 @@ public class KeysetAwareSliceImpl<T> implements KeysetAwareSlice<T> {
         }
 
         if (keysetPage == null) {
-            return new KeysetPageRequest(null, pageable.getSort(), pageable.getPageNumber(), pageable.getPageSize());
+            return new KeysetPageRequest(null, pageable.getSort(), pageable.getOffset(), pageable.getPageSize());
         } else {
             return new KeysetPageRequest(keysetPage, pageable.getSort());
         }