+++ integration/hibernate-4.2/src/main/java/com/blazebit/persistence/integration/hibernate/Hibernate4Integrator.java
@@ -81,7 +81,6 @@ public class Hibernate4Integrator implements ServiceContributingIntegrator {
             
             if (entityClass != null && entityClass.isAnnotationPresent(CTE.class)) {
                 clazz.getTable().setSubselect("select * from " + clazz.getJpaEntityName());
-                // TODO: check that no collections are mapped
             }
         }