+++ integration/hibernate-4.3/src/main/java/com/blazebit/persistence/integration/hibernate/Hibernate43Integrator.java
@@ -75,7 +75,6 @@ public class Hibernate43Integrator implements ServiceContributingIntegrator {
             
             if (entityClass != null && entityClass.isAnnotationPresent(CTE.class)) {
                 clazz.getTable().setSubselect("select * from " + clazz.getJpaEntityName());
-                // TODO: check that no collections are mapped
             }
         }